コントローラーの修正

+++ app/controllers/items_controller.rb
@@ -1,6 +1,6 @@
 class ItemsController < ApplicationController
   before_action :authenticate_user!, except: [:index, :show]
-  before_action :baria_user, only: [:edit, :update]
+  before_action :baria_user, only: [:edit, :update, :destroy]
   before_action :find_params, only: [:show, :edit, :update, :destroy]
 
   def index
@@ -35,11 +35,15 @@ class ItemsController < ApplicationController
   end
 
   def destroy
+    if current_user.id == @item.user.id
       if @item.destroy
         redirect_to root_path
       else
         render :show
       end
+    else
+      render :root_path
+    end
   end
